+++ src/main/java/kr/hhplus/be/server/interfaces/api/concert/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.request;
+package kr.hhplus.be.server.interfaces.api.concert.dto.request;
 
 import java.util.UUID;
