+++ src/main/java/kr/hhplus/be/server/interfaces/api/concert/dto/response/ConcertDateResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.response;
+package kr.hhplus.be.server.interfaces.api.concert.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
