+++ src/main/java/kr/hhplus/be/server/interfaces/api/concert/dto/response/QueueTokenResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.response;
+package kr.hhplus.be.server.interfaces.api.concert.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
