+++ src/main/java/kr/hhplus/be/server/interfaces/api/concert/dto/response/SeatResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.response;
+package kr.hhplus.be.server.interfaces.api.concert.dto.response;
 
 import java.math.BigDecimal;
 import java.util.UUID;
