+++ src/main/java/kr/hhplus/be/server/interfaces/api/user/UserController.java
@@ -17,8 +17,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.hhplus.be.server.interfaces.dto.request.ChargePointRequest;
-import kr.hhplus.be.server.interfaces.dto.response.UserPointResponse;
+import kr.hhplus.be.server.interfaces.api.user.dto.request.ChargePointRequest;
+import kr.hhplus.be.server.interfaces.api.user.dto.response.UserPointResponse;
 
 @RestController
 @RequestMapping("/api/v1/users")
