+++ src/main/java/kr/hhplus/be/server/interfaces/api/user/dto/response/UserPointResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.response;
+package kr.hhplus.be.server.interfaces.api.user.dto.response;
 
 import java.math.BigDecimal;
 import java.util.UUID;
